refactor: move utils helpers tests to tests folder
